@@ -1,6 +1,7 @@
 {
   "name": "travel-packing-app",
-  "version": "1.0.0",
+  "version": "0.1.0",
+  "private": true,
   "description": "A smart travel packing app",
   "main": "index.js",
   "scripts": {
@@ -13,13 +14,28 @@
     "node": "22.x"
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "^4.0.3",
-    "lucide-react": "^0.253.0"
+    "@types/node": "^16.18.0",
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
+    "ajv": "^8.12.0",
+    "ajv-keywords": "^5.1.0",
+    "lucide-react": "^0.253.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.9.5",
+    "web-vitals": "^2.1.4"
   },
   "devDependencies": {
-    "typescript": "^4.1.2"
+    "autoprefixer": "^10.4.20",
+    "postcss": "^8.5.3",
+    "tailwindcss": "^3.4.17"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
