@@ -1,5 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './travel-packing-app';
+import { createRoot } from 'react-dom/client';
+import './index.css';
+import PackingApp from './travel-packing-app';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const container = document.getElementById('root');
+if (!container) throw new Error('Failed to find the root element');
+const root = createRoot(container);
+
+root.render(
+  <React.StrictMode>
+    <PackingApp />
+  </React.StrictMode>
+);
